fixed namings & trailing spaces
